refactor(design): move input component outside of subfolder
