+++ libs/design/src/atoms/form/input/input.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../scss/daff-util';
+@import '../../../scss/daff-util';
 
 .daff-input {
 	border: 0;
