+++ libs/design/src/atoms/form/input/public_api.ts
@@ -1,2 +1,2 @@
 export { DaffInputModule } from './input.module';
-export * from './input/input.component';
+export { DaffInputComponent } from './input.component';
