+++ libs/design/src/scss/theming/_theme.scss
@@ -1,7 +1,7 @@
 @import '../../atoms/button/button-theme';
 @import '../../atoms/form/error-message/error-message-theme';
 @import '../../atoms/form/form-field/form-field/form-field-theme';
-@import '../../atoms/form/input/input/input-theme';
+@import '../../atoms/form/input/input-theme';
 @import '../../atoms/form/select/select/select-theme';
 @import '../../atoms/progress-indicator/progress-indicator-theme';
 @import '../../atoms/loading-icon/loading-icon-theme';
